refatorando os metodos para classe
