+++ Stopwatch/Main.cs
@@ -1,0 +1,10 @@
+﻿namespace Stopwatch
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Menu.menu();
+        }
+    }
+}
